requirements.txt
aiogram==3.3.0
asyncpg
python-dotenv
openai
